Try to place nonplaced pieces
If there are nonplaced pieces find the piece with the highest mutual
compatibility among them as first piece to place. Then chech where to place
it along the already placed pieces. Does not work really good so far.
Unfortunately pieces are tried to be placed at positions where there are
already pieces.