Changed folder structure of repo and added code for Ex1 AML
Everything from FoML is now in the folder with the corresponding name and
there is a new folder for AML where on can find the code for the exercices
for the lecture AML. FoML Project Code is still in its own folder. Will
maybe change that later on when Project is graded but for now I will leave it
as is so the link in the report that leads to the project's code will still
be correct.